app.nlp: moving functionality from object DocumentAnnotator to object DocumentAnnotatorPipeline
