+++ src/test/scala/cc/factorie/app/nlp/lexicon/TestLexicon.scala
@@ -35,7 +35,7 @@ class TestLexicon extends JUnitSuite {
     assert(!lexicon.containsWord("England"))
     
     val string = "Yesterday I flew from Paris to New York."
-    val doc = DocumentAnnotator.process(segment.RegexTokenizer, new Document(string))
+    val doc = DocumentAnnotatorPipeline.process(segment.RegexTokenizer, new Document(string))
     val section = doc.asSection
     assert(section.tokens(4).string == "Paris")
     assert(lexicon.contains(section.tokens(4)))
